refactor(custom_resources): move formatting textures part into scripts for faster startup times (#188)
