+++ src/lib/types/custom-resources.d.ts
@@ -83,3 +83,12 @@ export type OutputTexture = {
   debug: CustomResourcesDebugStats;
   skyblock_id: string;
 };
+
+interface ProgressInfo {
+  current: number;
+  total: number;
+  packName: string;
+  currentFile: string;
+  startTime: number;
+  errors: number;
+}
